refactor(hud): スクリーン座標を consts.Coord[consts.ScreenPixel] にする
HUD の AIStateInfo/VisionRangeInfo/HPDisplayInfo の Screen フィールドは、カメラ変換後の
画面座標なのに Coord[float64] でワールド座標と型が同じだった。Coord[consts.ScreenPixel] にして、
ワールド座標との取り違えを型で弾く。

抽出側の world→screen 計算は各所で重複していたので consts.WorldToScreen に集約する。
TileCenterToWorld と合わせ、グリッド→ワールドピクセル→スクリーンピクセルの変換段が型で表れる。
描画境界では ebiten の float64/float32 API へキャストする。挙動は不変。

+++ internal/widgets/hud/debug_overlay.go
@@ -39,7 +39,7 @@ func (overlay *DebugOverlay) Draw(screen *ebiten.Image, data DebugOverlayData) {
 	// AI状態を描画
 	for _, aiState := range data.AIStates {
 		textOffsetY := 30.0
-		drawOutlinedText(screen, aiState.StateText, overlay.face, float64(int(aiState.Screen.X)-20), aiState.Screen.Y-textOffsetY, theme.TextPrimary)
+		drawOutlinedText(screen, aiState.StateText, overlay.face, float64(int(aiState.Screen.X)-20), float64(aiState.Screen.Y)-textOffsetY, theme.TextPrimary)
 	}
 
 	// 視界範囲を描画
@@ -51,7 +51,7 @@ func (overlay *DebugOverlay) Draw(screen *ebiten.Image, data DebugOverlayData) {
 	for _, hpDisplay := range data.HPDisplays {
 		hpText := fmt.Sprintf("%d/%d", hpDisplay.CurrentHP, hpDisplay.MaxHP)
 		textOffsetY := 15.0 // AI状態テキスト（30.0）より上に表示して重複を避ける
-		drawOutlinedText(screen, hpText, overlay.face, float64(int(hpDisplay.Screen.X)-15), hpDisplay.Screen.Y-textOffsetY, theme.TextPrimary)
+		drawOutlinedText(screen, hpText, overlay.face, float64(int(hpDisplay.Screen.X)-15), float64(hpDisplay.Screen.Y)-textOffsetY, theme.TextPrimary)
 	}
 }
